Updates API server concerns for relocated networking types.
