+++ apiserver/facades/client/applicationoffers/state.go
@@ -10,7 +10,7 @@ import (
 
 	commoncrossmodel "github.com/juju/juju/apiserver/common/crossmodel"
 	"github.com/juju/juju/core/crossmodel"
-	"github.com/juju/juju/network"
+	"github.com/juju/juju/core/network"
 	"github.com/juju/juju/permission"
 	"github.com/juju/juju/state"
 )
